Cache, error handling, string formatting
Cache: added an initialized cache variable at the beginning of the program. This variable is used to store the ISO2 codes of the countries that have already been converted so that they do not have to be reconverted each time.

Error handling: modified the getflag function so that it raises a ValueError exception if it cannot find the ISO2 code for a given country name.

Flexibility in string formatting: modified the main function so that it accepts country names separated by spaces, commas, or semicolons.